tweaks and bug fixes
changed Nerd outfit to increase quantity of research found in loot.
extended the reach on the paintbrush.
legendary items are now found in POI end-loot in the Wasteland only. The higher the POI tier, the more likely they are to be found.
Increased research papers found in zombie corpses in the wasteland.
Lowered loot probability for hunting weapons from lumberjacks.
Fixed challenge rewards for cop and mutated zombies.
Hopefully fixed the sledgehammer 'infinite stamina' bug.
Localization tweaks and fixes.